[chore] relocate input common components
