+++ src/features/plan/input/index.ts
@@ -1,0 +1,2 @@
+export { default as DataCategoriesSection } from "./widgets/dataCategoriesSection";
+export { default as BreadcrumbSection } from "./widgets/breadcrumbSection";
